fix(toolset-server): drain in-flight requests before deregistering
On SIGTERM, serve_with_shutdown deleted the KV registration and
returned without waiting for in-flight tool requests. A caller
already waiting on a reply for one of those got nothing back and
blocked until its own 60s timeout -- the same class of hang as the
stale-registration issue fixed earlier, just triggered by an
in-flight call instead of a leftover registry entry.

Chose to drain rather than fail calls explicitly: aborting in-flight
tasks would turn a delayed reply into no reply at all, which is worse
for the caller. Track active requests with a watch-based counter (an
RAII guard increments on spawn, decrements on drop so a panic can't
leak the count) and wait for it to reach zero before deregistering,
bounded by a 10s deadline so a stuck invocation can't stall shutdown
forever. Requests still running past the deadline keep running and
still reply -- only the wait itself is bounded.

Extracted the counter and drain logic into a new drain module: it
shares no data with the idempotency cache or registration-refresh
code, and keeping it in lib.rs pushed that file over CodeScene's
cohesion threshold (a new issue on the first attempt). Split the
integration test file's nats-server/harness plumbing into
tests/common/mod.rs for the same reason on the test side.

Added a regression test using an echo call with a delay (not the
existing 'slow' tool + cancel message, since once shutdown exits the
main select loop it stops polling the control subscription that
cancel relies on). Confirmed fail-then-pass: without the drain, the
test fails on 'registration must still be present while the request
is in flight'.
